requirements.txt
langchain
langchain_community
ipykernel
python-dotenv
langchain_ollama
langchain_groq
streamlit

